Refactor generic method into mixin class
